Put domain classes in their own domain folders
At first I thought it's a good idea to put these domain classes as a
top-level of the slice. But with that, they get mixed up with technical
concerns, such a Slice::Action or authorization. Instead, let's now put
them in their respective "domain" directories, which should improve discoverability.
